changed inventory hashmaps to use <Long, Integer>

+++ src/main/java/inventorycount/InventoryCountController.java
@@ -1,6 +1,5 @@
 package inventorycount;
 
-import javax.naming.InsufficientResourcesException;
 import java.util.HashMap;
 import java.util.UUID;
 
@@ -15,14 +14,14 @@ public class InventoryCountController {
     }
 
 
-    public void submitCount(HashMap<Integer, Integer> count){
+    public void submitCount(HashMap<Long, Integer> count){
 
         this.inventoryCountInteractor.updateInventoryCount(count);
         // TODO: return logic for presenter, void for now
 
     }
 
-    public HashMap<Integer, Integer> getCurrentCount(){
+    public HashMap<Long, Integer> getCurrentCount(){
         return inventoryCountInteractor.getCurrentCount();
     }
